fix lint:eslint issues
