+++ src/env.d.ts
@@ -1,2 +1,2 @@
-/// <reference path="../.astro/types.d.ts" />
+import '../.astro/types.d.ts';
 /// <reference types="astro/client" />
