+++ tsconfig.json
@@ -1,9 +1,12 @@
 {
   "extends": "astro/tsconfigs/strictest",
+  "include": [".astro/types.d.ts", "**/*.ts", "**/*.tsx", "**/*.astro"],
+  "exclude": ["node_modules", "dist"],
   "compilerOptions": {
     "jsx": "react-jsx",
     "allowJs": true,
     "checkJs": true,
+    "strict": true,
     "baseUrl": ".",
     "paths": {
       "~/*": ["src/*"]
